// client/src/app/_models/hmi.ts
export class View {
    /** View id, random number */
    id = '';
    /** View name used as reference in configuration */
    name = '';
    /** View profile size, background color */
    profile: DocProfile = new DocProfile();
    /** Gauges settings list used in the view  */
    items: DictionaryGaugeSettings = {};
    /** Variables (Tags) list used in the view */
    variables: DictionaryVariables = {};
    /** Svg code content of the view  */
    svgcontent = '';
    /** Type of view SVG/CARDS */
    type: ViewType;
    /** Property with events of view like Open or Close */
    property: ViewProperty;
    /** Folder id for organizing views, null/undefined = root level */
    folderId?: string;

    constructor(id?: string, type?: ViewType, name?: string) {
        this.id = id;
        this.name = name;
        this.type = type;
    }
}

export enum ViewType {
    svg = 'svg',
    cards = 'cards',
    maps = 'maps'
}
// ...
export class DocProfile {
    width = 1024;
    height = 768;
    bkcolor = '#ffffffff';
    margin = 10;
    align = DocAlignType.topCenter;
    gridType: GridType = GridType.Fixed;
    viewRenderDelay = 0; // delay in ms to render view after load, used to prevent flickering on load view with many gauges
}

export enum DocAlignType {
    topCenter = 'topCenter',
    middleCenter ='middleCenter'
}

export class GaugeSettings {
    name = '';
    property: any = null;   // set to GaugeProperty after upgrate
    label = '';             // Gauge type label
    hide = false;
    lock = false;
    constructor(public id: string, public type: string) {
    }
}

export class ViewProperty {
    events: GaugeEvent[] = [];
    startLocation?: MapsLocation;
    startZoom?: number;
}
export class GaugeProperty {
    variableId: string;
    variableValue: string;
    bitmask: number;
    permission: number;
    permissionRoles: PermissionRoles;
    ranges: GaugeRangeProperty[];
    events: GaugeEvent[] = [];
    actions: GaugeAction[] = [];
    options: any;
    readonly: boolean;
    text: string;               // Text property (used by button)
}

export interface PermissionRoles {
    show: string[];
    enabled: string[];
}

export class WidgetProperty extends GaugeProperty {
    type: string;
    scriptContent?: { moduleId: string, content: string };
    svgContent?: string;
    varsToBind?: WidgetPropertyVariable[] = [];
}
// ...
export class GaugeEvent {
    type: string;
    action: string;
    actparam: string;
    actoptions = <any>{};
}
// ...
export class GaugeAction {
    variableId: string;
    bitmask: number;
    range: GaugeRangeProperty;
    type: any;
    options: any = {};
}
// ...
export class GaugeRangeProperty {
    min: number;
    max: number;
    minVariableId?: string;
    maxVariableId?: string;
    text: string;
    textId: string;
    color: string;
    type: any;
    style: any;
    stroke: string;
    address?: string;
}
// ...
export class Variable {
    id: string;
    name: string;
    source: string;
    value: string;
    error: number;
    timestamp: number;
    device?: Device;
    quality?: string;
    constructor(id: string, name: string, device?: Device) {
        this.id = id;
        this.name = name;
        this.device = device;
        if (device?.type === DeviceType.internal) {
            this.value = '0';
        }
    }
}
// ...
export interface DictionaryGaugeSettings {
    [x: string]: GaugeSettings;
}

export interface DictionaryVariables {
    [id: string]: Variable;
}